+++ tests/test_shebangs.py
@@ -3,9 +3,10 @@ import stat
 from typing import Optional
 
 import git_file_utils
+from get_repo_root import get_repo_root
 
 
-REPO_ROOT = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
+REPO_ROOT = get_repo_root()
 SKIP_DIRS = {
 	".git",
 	".venv",